Move insert quotes transaction filter to separate file
